move: 패키지 구조 세분화
application, domain, infrastructure 패키지 생성 후 재분류
- application: dto, exception
- domain: entity
- infrastructure: config, util

Resolved #41
