+++ common/src/main/java/com/business/common/application/exception/ExceptionCode.java
@@ -1,4 +1,4 @@
-package com.business.common.exception;
+package com.business.common.application.exception;
 
 import org.springframework.http.HttpStatusCode;
