+++ common/src/main/java/com/business/common/domain/entity/BaseDataEntity.java
@@ -1,6 +1,6 @@
-package com.business.common.entity;
+package com.business.common.domain.entity;
 
-import com.business.common.util.CommonUtil;
+import com.business.common.infrastructure.util.CommonUtil;
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
@@ -9,7 +9,6 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Getter
 @MappedSuperclass
@@ -17,40 +16,40 @@ import java.util.UUID;
 public abstract class BaseDataEntity {
 
     @Column(name = "created_by", nullable = false, updatable = false)
-    private UUID createdBy;
+    private Long createdBy;
 
     @CreatedDate
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
 
     @Column(name = "updated_by")
-    private UUID updatedBy;
+    private Long updatedBy;
 
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
     @Column(name = "deleted_by")
-    private UUID deletedBy;
+    private Long deletedBy;
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
 
-    public void setCreatedBy(UUID userId) {
+    public void setCreatedBy(Long userId) {
         this.createdBy = userId;
         this.createdAt = LocalDateTime.now();
     }
 
-    public void setUpdatedBy(UUID userId) {
+    public void setUpdatedBy(Long userId) {
         this.updatedBy = userId;
         this.updatedAt = LocalDateTime.now();
     }
 
-    public void setDeletedBy(UUID userId) {
+    public void setDeletedBy(Long userId) {
         this.deletedBy = userId;
         this.deletedAt = LocalDateTime.now();
     }
 
-    public void initBaseDataFromToken(UUID userId) {
+    public void initBaseDataFromToken(Long userId) {
         String nowAsString = CommonUtil.LDTToString(LocalDateTime.now());
         this.createdBy = userId;
         this.createdAt = CommonUtil.stringToLDT(nowAsString);
