+++ common/src/main/java/com/business/common/infrastructure/util/QueryDslUtil.java
@@ -1,4 +1,4 @@
-package com.business.common.util;
+package com.business.common.infrastructure.util;
 
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
