+++ common/src/test/java/com/business/common/application/exception/MockExceptionCode.java
@@ -1,4 +1,4 @@
-package com.business.common.exception;
+package com.business.common.application.exception;
 
 import org.springframework.http.HttpStatus;
