+++ common/src/test/java/com/business/common/infrastructure/util/CommonUtilTest.java
@@ -1,4 +1,4 @@
-package com.business.common.util;
+package com.business.common.infrastructure.util;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.jupiter.api.Test;
